♻️ refactor: auth api, queries 리팩토링
